updated the seeder file updated the projectController to handle the tags corrected usersSvc by adding the  @service made updates to the pom file
