+++ src/main/java/com/artiscene/services/usersSvc.java
@@ -2,13 +2,15 @@ package com.artiscene.services;
 
 import com.artiscene.models.Project;
 import com.artiscene.models.User;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
 
 /**
  * Created by vanessamnoble on 2/28/17.
  */
-
+@Service("usersSvc")
 public class usersSvc {
 
     public boolean isLoggedIn() {
@@ -22,9 +24,5 @@ public class usersSvc {
             return null;
         }
         return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-    }
-
-    public void save(Project projectCreated) {
-
     }
 }
