+++ tagSeeder.sql
@@ -1,10 +1,11 @@
 USE artiscene_db;
+
 INSERT into tags (tag_name) VALUES ("acrylic");
 INSERT into tags (tag_name) VALUES ("watercolor");
 INSERT into tags (tag_name) VALUES ("charcoal");
 INSERT into tags (tag_name) VALUES ("ink");
 INSERT into tags (tag_name) VALUES ("color");
-INSERT into tags (tag_name) VALUES ("blackandwhite");
+INSERT into tags (tag_name) VALUES ("black and white");
 INSERT into tags (tag_name) VALUES ("mixed");
 INSERT into tags (tag_name) VALUES ("prints");
 INSERT into tags (tag_name) VALUES ("woodcut");
@@ -20,7 +21,7 @@ INSERT into tags (tag_name) VALUES ("oil");
 INSERT into tags (tag_name) VALUES ("tempera");
 INSERT into tags (tag_name) VALUES ("pastel");
 INSERT into tags (tag_name) VALUES ("fabric");
-INSERT into tags (tag_name) VALUES ("scotchboard");
+INSERT into tags (tag_name) VALUES ("scratchboard");
 INSERT into tags (tag_name) VALUES ("wood");
 INSERT into tags (tag_name) VALUES ("wire");
 INSERT into tags (tag_name) VALUES ("glass");
@@ -33,7 +34,7 @@ INSERT into tags (tag_name) VALUES ("enamel");
 INSERT into tags (tag_name) VALUES ("ink wash");
 INSERT into tags (tag_name) VALUES ("stippling");
 INSERT into tags (tag_name) VALUES ("line work");
-INSERT into tags (tag_name) VALUES ("natching");
+INSERT into tags (tag_name) VALUES ("hatching");
 INSERT into tags (tag_name) VALUES ("traditional");
 INSERT into tags (tag_name) VALUES ("mosaic");
 INSERT into tags (tag_name) VALUES ("photography");
@@ -52,14 +53,14 @@ INSERT into tags (tag_name) VALUES ("commissioned");
 INSERT into tags (tag_name) VALUES ("w.i.p");
 INSERT into tags (tag_name) VALUES ("graphic design");
 INSERT into tags (tag_name) VALUES ("performance");
-INSERT into tags (tag_name) VALUES ("phototshop");
+INSERT into tags (tag_name) VALUES ("photoshop");
 INSERT into tags (tag_name) VALUES ("collage");
-INSERT into tags (tag_name) VALUES ("amature");
+INSERT into tags (tag_name) VALUES ("amatur");
 INSERT into tags (tag_name) VALUES ("professional");
 INSERT into tags (tag_name) VALUES ("graffiti");
 INSERT into tags (tag_name) VALUES ("mural");
 INSERT into tags (tag_name) VALUES ("street art");
-INSERT into tags (tag_name) VALUES ("cryon");
+INSERT into tags (tag_name) VALUES ("crayon");
 INSERT into tags (tag_name) VALUES ("sand");
 INSERT into tags (tag_name) VALUES ("color pencil");
 INSERT into tags (tag_name) VALUES ("figure drawing");
